[CLI] Move cli code into cli/, because more nesting never hurt anyone
