+++ src/cli/watch.js
